GET /albums/:albumId/products/:productId get a product for an album refactored product services to ProductsApi
